Assorted gql performance improvements
Summary:
Change Struct to be a C-style interface. It has a concrete vtable ptr, and Struct implementations do explicit unsafe casts.
This achieves the same effect as Go interfaces, but it saves a few malloc and pointer indirections. This saves about 10% overhead e2e in a benchmark.

Add a templetized thread-safe freepool. This was originally used in sam RecordPool, but this change it moves to base/gtl and instantiates in a few places in gql. This saves  about 20% overhead.

Reduce the size of Value from 32 bytes to 24 bytes.

Reviewers: marius, pbohannon

Reviewed By: marius

Differential Revision: https://phabricator.grailbio.com/D15267

fbshipit-source-id: 672c127