github.com/grailbio/base/simd: byte-interleave and reverse
Summary:
This adds functions for interleaving two byte-slices, and reversing a
slice in-place and not-in-place.  (The latter functions are very
similar to the reverse-complement functions I have planned for
bio/biosimd.)

These are the last major functions in my current base/simd design; work
will now start on bio/biosimd.

Test Plan: Unit tests and benchmarks have been added.

Reviewers: ysaito, ayip

Reviewed By: ysaito

Subscribers: cchang

Differential Revision: https://phabricator.grailbio.com/D15581

fbshipit-source-id: 4338a0b