bio-spamtools: add -region/-ibed options to kmer-repeats
Summary:
This extends the kmer-repeats command so that it can consider only
reads mapped to the given set of intervals.  (The sharding strategy in
this case is pretty crude.  We probably want to add a better way to
handle this to bamprovider in the future.)

In the process, a BED reader bug involving improper use of gunsafe was
fixed.

(It also adds documentation to simd.Memset8 to the effect that it
shouldn't be used for val == 0, with benchmark results; and adds a few
uses of //go:linkname to reduce redundancy between base/simd and
bio/biosimd.)

Test Plan: Output matches C-spamtool; adding an integration test for this now.

Reviewers: ysaito, ayip

Reviewed By: ysaito

Subscribers: cchang

Differential Revision: https://phabricator.grailbio.com/D16206

fbshipit-source-id: 7e144d1