bio/bagsampler: first working version
Summary:
This now (reproducibly) samples bags at the given rate and produces
pileup-like output for them.

Test Plan:
Unit test for newCigarIteratorPair() is included.  Ran this multiple
times on a 190 GB dupmarked BAM with the same sampling rate
and random seed (on a 40-core adhoc instance), and confirmed
that the output was identical after "sort -k1,1 -s".

Integration test is being postponed to the next diff, which will
split this into bagsampler and bagreporter binaries.

Reviewers: ayip, ysaito

Reviewed By: ysaito

Subscribers: escott, psingh, dburkhardt, avalouev, cchang

Differential Revision: https://phabricator.grailbio.com/D19370

fbshipit-source-id: 764eda1