bio/encoding/bam: add .gbai index file format for .bam files
Summary:
This includes a new command line tool to index a .bam file, and adds
support to bamprovider for .gbai files.  To use it, just specify a
.gbai file in place of a .bai file.

The shard sizes for a .gbai file are substantially more even than
those from a .bai file, and this helps to reduce stragglers.

This code change does not split a given genomic position into multiple
shards, nor does it split the unmapped region into multiple shards.
The .gbai index format does however allow for future support in the
same way that PAM does using sequence numbers.

In my test 14GB file, this results in 11% shorter runtime for bio-mark-dups.
Time and memory reduction should be better with targetted input files.

Test Plan:
Added unit tests.
Ran markdups with a 14GB file with .bai and .gbai files and compared outputs to be identical.

Reviewers: cchang, ysaito

Reviewed By: cchang, ysaito

Differential Revision: https://phabricator.grailbio.com/D17215

fbshipit-source-id: 787b4d0