bio/encoding/bammates: separate distant mates from bio-mark-duplicates
Summary:
This change separates the distant mates functionality from bio-mark-duplicates so that we could use it in other code.

We'll use this in a later change to make PairIterator deterministic which is related to the task of making our pipelines completely deterministic: T12808

We can also use this to make bio-metrics operate on a per-shard basis so that it need not store the entire coverage map in memory, which is related to fixit task T13304

Fixes T12808
Fixes T13304

Test Plan: Preserved existing unit tests, and added a couple more test cases.

Reviewers: ysaito

Reviewed By: ysaito

Maniphest Tasks: T13304, T12808

Differential Revision: https://phabricator.grailbio.com/D21704

fbshipit-source-id: a44a763