bio/cmd/bio-bam-sort: make sorting determistic
Summary:
The prior version of bio-bam-sort uses (ref, position, direction) as a sort key, and relies to sort stability to be deterministic.  This is not sufficient when the inputs arrive in non-deterministic order, and siege outputs records in non-deterministic order.

This change makes bio-bam-sort use the serialized bam record as a tiebreaker when records collide on (ref, position, direction).  The serialized encoding created by our bam.Marshal() is deterministic for a given record's content, so using the serialized encoding as a key is safe.  The only caveat is that serialized encoding does not create a canonical order for the aux tags, so two records with identical but differently ordered aux tags would be sorted to different locations in the output.

Test Plan:
Ran current bazel tests.

Removed some test cases that don't make sense anymore.

Reviewers: ysaito, escott, sbanville

Reviewed By: ysaito

Subscribers: psingh, escott, sbanville, ysaito, ayip

Tags: #methylation_pipeline

Maniphest Tasks: T15677

Differential Revision: https://phabricator.grailbio.com/D24885

fbshipit-source-id: 23965ec