bio-mark-duplicates: fix bug in padding check
Summary:
The previous version checks just the clipping distance on either end
of a read to check that the shards have enough padding. This is
insufficient because on a reverse read, the alignment position is on
the left side of the read, and the 5' is on the left side of the read.
In this reverse-read case, the distance between the alignment position
and the unclipped 5' end of the read can be greater than the right
clipping distance of such a read.  For example, if a reverse read has
the cigar 10M5D10M, the distance from the alignment position to the 5'
end is 24, even though there is no clipping. The shard padding in this
case must be at least 24, and the previous version of the code was not
checking this distance of 24 against the user's chosen padding value.

When a read has a distance that exceeds the padding, then
bio-mark-duplicates returns incorrect duplicate flags and also
incorrect DI and DS values.

This change adds the corrected distance calculation and a check that
verifies the distance against the user's chosen padding value.

This change also bubbles up errors that occur during this check to
facilitate unit tests.

Test Plan:
Add unit tests for alignment distance check and its integration with
mark-duplicates.  Patch up tests to work with new alignment distance
check.

Reviewers: akokate, escott

Reviewed By: akokate

Maniphest Tasks: T20864

Differential Revision: https://phabricator.grailbio.com/D31099

fbshipit-source-id: 842377f