grailbio/bio/interval: distinguish intervals from interval-endpoints
Summary:
This defines a small interval-endpoints API (and modifies existing code
to use it when appropriate), and defines the internal
disjointIntervalSet type for the data structures that should not be
locked to an interval-endpoints representation.

(An update to the getTokens() utility function is also included in
this diff.  That can be moved into its own diff... preferably one
which moves it to an appropriate public location.)

Ref T20149

Test Plan: Existing unit tests pass.

Reviewers: ayip, ysaito

Reviewed By: ayip

Subscribers: cchang

Maniphest Tasks: T20149

Differential Revision: https://phabricator.grailbio.com/D31355

fbshipit-source-id: 19c7f78