bio/biosimd: write linefeeds and '+' correctly
Summary:
The first test of the full demuxer revealed this mistake (wrote to a
register instead of the address the register pointed to).

Ref T20646

Test Plan:
Unit test has been extended to validate this part of the record.
Remaining differences between decompressed demuxer fastq output and
the expected value on a test dataset appear to be unrelated to this
function.

Reviewers: ysaito

Reviewed By: ysaito

Subscribers: cchang

Maniphest Tasks: T20646

Differential Revision: https://phabricator.grailbio.com/D32347

fbshipit-source-id: 0b17983