cmd/bio-pileup: command-line parsing skeleton
Summary:
This creates a skeleton bio-pileup binary which has the intended
command-line interface.  Actual pileup logic will be placed under a
snp/ subdirectory in the next diff.

Ref T20121
Ref T20149

Test Plan:
Verified that -help spells out the positional arguments, and that an
error message is printed if the wrong number of positional arguments
is given.

Reviewers: ayip

Reviewed By: ayip

Subscribers: osakarya, cchang

Maniphest Tasks: T20149, T20121

Differential Revision: https://phabricator.grailbio.com/D44920

fbshipit-source-id: 2b718cf