bio/encoding/bam: fix gbai byte-based sharding
Summary:
During testing of D46458, the gbai byte-based sharder mysteriously
returned only 5 shards when targetParallelism was 40.  This was caused
by a minBases check that didn't take reference differences into
account.

This diff fixes the minBases check.

Test Plan:
Existing tests pass.  Verified that with this change and D46458,
bio-collapse performance on a 14 GB test BAM is at least as good as it
is with .bai (which was not true before this change), and results are
correct.

Reviewers: ayip

Reviewed By: ayip

Subscribers: cchang

Maniphest Tasks: T38116

Differential Revision: https://phabricator.grailbio.com/D46554

fbshipit-source-id: 4da5567