bio/encoding/fasta: fixit: add support for BAM-friendly fasta load
Summary:
When working with BAM-formatted data, it can be more convenient to
have the reference genome encoded in the BAM binary format (A=1, C=2,
G=4, T=8, N=15) instead of ASCII.  This was previously handled by
ad-hoc bio-pileup code, and was the reason it did not use
bio/encoding/fasta.

Test Plan:
Existing tests pass, along with an appropriately augmented version of
the biosimd ASCIIToSeq8 test.  Further testing will come with the next
​diff, which will revise bio-pileup, bio-mismatch-tag, bio-spamtools,
and conta-validation to use bio/encoding/fasta.

Reviewers: joshnewman

Reviewed By: joshnewman

Subscribers: ghorrell, escott, ayip, sbagaria, cchang

Maniphest Tasks: T32685

Differential Revision: https://phabricator.grailbio.com/D46599

fbshipit-source-id: 2ddabba